% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba.R
\docType{data}
\name{nba}
\alias{nba}
\title{NBA Player Statistics (1950-2017)}
\format{
A data frame with 24691 rows and 52 variables:
\describe{
\item{Year}{Year of that NBA season}
\item{Player}{Player's name}
\item{Age}{Player's age at that time, in years}
\item{G}{Number of games played by the specified player in that NBA season}
}
}
\source{
\url{https://www.kaggle.com/drgilermo/nba-players-stats/data?select=Seasons_Stats.csv}
}
\usage{
nba
}
\description{
The data-set contains aggregate individual statistics for 3,922 unique players over 67 NBA seasons, from basic box-score attributes such as points, assists, rebounds etc., to more advanced money-ball like features such as Value Over Replacement.
}
\keyword{datasets}
