% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_meangames.R
\name{year_meangames}
\alias{year_meangames}
\title{Mean Games per Player}
\usage{
year_meangames(x)
}
\arguments{
\item{year}{Year of interest}
}
\value{
Output is the Mean number of games played by each player in that year
}
\description{
This function calculates the mean number of games played per player, in that particular NBA year. Imports: dplyr
}
\examples{
year_meangames(2011)
}
