% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_numeric.R
\name{year_numeric}
\alias{year_numeric}
\title{Correlation Matrix}
\usage{
year_numeric(x)
}
\arguments{
\item{year}{Year of interest}
}
\value{
Output is a correlation matrix
}
\description{
This function calculates the correlation matrix of numeric variables, in that particular NBA year. Imports: dplyr
}
\examples{
year_numeric(2011)
}
