% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_oldestplayer.R
\name{year_oldestplayer}
\alias{year_oldestplayer}
\title{Oldest player in that year}
\usage{
year_oldestplayer(x)
}
\arguments{
\item{year}{Year of interest}
}
\value{
Output is a one-row tibble displaying the name of the oldest player and their age (in that year).
}
\description{
This function retrieves the player with the oldest age, in that particular NBA year. Imports: dplyr
}
\examples{
year_oldestplayer(2011)
}
